Refactored Thumbor into its own module, next to RPAPI
