+++ app/templates/index.html
@@ -2,50 +2,16 @@
 
 {% set page_title = 'Homepage' %}
 
-{% block custom_css %}
-    <style>
-        .example {
-            background-color: aqua;
-        }
-    </style>
-{% endblock %}
-
-{% block custom_js %}
-    <script>
-        $(document).ready(function () {
-            $("#submit").bind("click", function(){
-                var urlpath = $("#path").val();
-                var input_data = {'path': urlpath};
-                var cacheurl = "{{ url_for('HomeView:post') }}";
-                $.post(cacheurl, input_data, function (data) {
-                    $("#output").text(decodeURI(data));
-                });
-            });
-
-            // Ensure the path contains a '/' in the front and no spaces
-            $('#path').keyup(function () {
-                var input = $(this);
-                input.val(input.val().replace(/ /g, ''));
-                var value = input.val();
-                if (value[0] != '/' && value.length > 0) {
-                    input.val("/" + value);
-                    value = input.val();
-                }
-            });
-        });
-    </script>
-{% endblock %}
-
 {% block main_content %}
     <div class="content">
-        <div class="form">
-            <h4>Thumbor Cache</h4>
-            <label>Path
-                <input type="text" name="path" id="path"/>
-            </label>
-            <a id="submit" class="btn btn-primary">Submit</a>
-        </div>
-        <p id="output" style="white-space: pre"></p>
+        <ul>
+            <li>
+                <a href="{{ url_for('ThumborView:index') }}">Clear Thumbor's caches</a>
+            </li>
+            <li>
+                <a href="{{ url_for('RefreshPurgeView:index') }}">Use the Refresh/Purge API</a>
+            </li>
+        </ul>
     </div>
 {% endblock %}
