+++ app/templates/thumbor.html
@@ -1,0 +1,48 @@
+{% extends 'bases/base.html' %}
+
+{% set page_title = 'Thubmor' %}
+
+{% block main_content %}
+    <div class="content">
+        <div class="form">
+            <h4>Thumbor Cache</h4>
+            <label>Path
+                <input type="text" name="path" id="path"/>
+            </label>
+            <a id="submit" class="btn btn-primary">Submit</a>
+        </div>
+        <p id="output" style="white-space: pre"></p>
+    </div>
+{% endblock %}
+
+{% block custom_js %}
+    <script>
+        $(document).ready(function () {
+            $("#submit").bind("click", function(){
+                var urlpath = $("#path").val();
+                var input_data = {'path': urlpath};
+                var cacheurl = "{{ url_for('ThumborView:submit') }}";
+                $.post(cacheurl, input_data, function (data) {
+                    $("#output").text(decodeURI(data));
+                });
+            });
+
+            // Ensure the path contains a '/' in the front and no spaces
+            $('#path').keyup(function () {
+                var input = $(this);
+                input.val(input.val().replace(/ /g, ''));
+                var value = input.val();
+                if (value.length > 0 && value[0] !== '/') {
+                    input.val('/' + value);
+                    // value = input.val();
+                }
+            });
+        });
+    </script>
+{% endblock %}
+
+
+
+
+
+
