+++ app/templates/thumbor_success.html
@@ -1,4 +1,4 @@
 {% if matches and cleared_path %}
 {{ matches|safe }}
-Cache cleared for path, "www.bethel.edu{{ cleared_path }}".
+Cache cleared for path "{{ cleared_path }}".
 {% endif %}
