+++ app/views/rpapi.py
@@ -1,8 +1,11 @@
+# Imports from Python global packages
 import re
 
+# Imports from packages installed from requirements.txt
 from flask import render_template, request
 from flask_classy import FlaskView, route
 
+# Imports from elsewhere in this project
 from app.forms.rpapi import PurgeRefreshForm, SimpleBanForm, AdvancedBanForm
 from app.controllers.rpapi import rpapi_call
